+++ services/galley/src/Galley/API/Util.hs
@@ -62,6 +62,7 @@ import Network.Wai.Predicate hiding (Error)
 import Network.Wai.Utilities
 import UnliftIO (concurrently)
 import qualified Wire.API.Conversation as Public
+import Wire.API.ErrorDescription
 import qualified Wire.API.Federation.API.Brig as FederatedBrig
 import Wire.API.Federation.API.Galley as FederatedGalley
 import Wire.API.Federation.Client (FederationClientFailure, FederatorClient, executeFederated)
@@ -73,14 +74,14 @@ type JSON = Media "application" "json"
 
 ensureAccessRole :: AccessRole -> [(UserId, Maybe TeamMember)] -> Galley ()
 ensureAccessRole role users = case role of
-  PrivateAccessRole -> throwM convAccessDenied
+  PrivateAccessRole -> throwErrorDescription convAccessDenied
   TeamAccessRole ->
     when (any (isNothing . snd) users) $
-      throwM notATeamMember
+      throwErrorDescription notATeamMember
   ActivatedAccessRole -> do
     activated <- lookupActivatedUsers $ map fst users
     when (length activated /= length users) $
-      throwM convAccessDenied
+      throwErrorDescription convAccessDenied
   NonActivatedAccessRole -> return ()
 
 -- | Check that the given user is either part of the same team(s) as the other
@@ -117,7 +118,7 @@ ensureConnectedToLocals u uids = do
     getConnections [u] (Just uids) (Just Accepted)
       `concurrently` getConnections uids (Just [u]) (Just Accepted)
   unless (length connsFrom == length uids && length connsTo == length uids) $
-    throwM notConnected
+    throwErrorDescription notConnected
 
 ensureReAuthorised :: UserId -> Maybe PlainTextPassword -> Galley ()
 ensureReAuthorised u secret = do
@@ -132,7 +133,7 @@ ensureReAuthorised u secret = do
 ensureActionAllowed :: Action -> InternalMember a -> Galley ()
 ensureActionAllowed action mem = case isActionAllowed action (memConvRoleName mem) of
   Just True -> return ()
-  Just False -> throwM (actionDenied action)
+  Just False -> throwErrorDescription (actionDenied action)
   Nothing -> throwM (badRequest "Custom roles not supported")
 
 -- Actually, this will "never" happen due to the
@@ -159,8 +160,8 @@ permissionCheck p = \case
   Just m -> do
     if m `hasPermission` p
       then pure m
-      else throwM (operationDenied p)
-  Nothing -> throwM notATeamMember
+      else throwErrorDescription (operationDenied p)
+  Nothing -> throwErrorDescription notATeamMember
 
 assertTeamExists :: TeamId -> Galley ()
 assertTeamExists tid = do
@@ -172,7 +173,7 @@ assertTeamExists tid = do
 assertOnTeam :: UserId -> TeamId -> Galley ()
 assertOnTeam uid tid = do
   Data.teamMember tid uid >>= \case
-    Nothing -> throwM notATeamMember
+    Nothing -> throwErrorDescription notATeamMember
     Just _ -> return ()
 
 -- | If the conversation is in a team, throw iff zusr is a team member and does not have named
@@ -183,7 +184,7 @@ permissionCheckTeamConv zusr cnv perm =
     Just cnv' -> case Data.convTeam cnv' of
       Just tid -> void $ permissionCheck perm =<< Data.teamMember tid zusr
       Nothing -> pure ()
-    Nothing -> throwM convNotFound
+    Nothing -> throwErrorDescription convNotFound
 
 -- | Try to accept a 1-1 conversation, promoting connect conversations as appropriate.
 acceptOne2One :: UserId -> Data.Conversation -> Maybe ConnId -> Galley Data.Conversation
@@ -199,7 +200,7 @@ acceptOne2One usr conv conn = do
           return $ conv {Data.convLocalMembers = mems <> toList mm}
     ConnectConv -> case mems of
       [_, _] | usr `isMember` mems -> promote
-      [_, _] -> throwM convNotFound
+      [_, _] -> throwErrorDescription convNotFound
       _ -> do
         when (length mems > 2) $
           throwM badConvState
@@ -267,7 +268,7 @@ membersToRecipients (Just u) = map userRecipient . filter (/= u) . map (view use
 -- the conversation, we don't want to disclose that such a conversation
 -- with that id exists.
 getSelfMember :: Foldable t => UserId -> t LocalMember -> Galley LocalMember
-getSelfMember = getMember convNotFound
+getSelfMember = getMember (errorDescriptionToWai convNotFound)
 
 getOtherMember :: Foldable t => UserId -> t LocalMember -> Galley LocalMember
 getOtherMember = getMember convMemberNotFound
@@ -281,14 +282,16 @@ getMember ex u ms = do
     Nothing -> throwM ex
 
 getConversationAndCheckMembership :: UserId -> ConvId -> Galley Data.Conversation
-getConversationAndCheckMembership = getConversationAndCheckMembershipWithError convAccessDenied
+getConversationAndCheckMembership =
+  getConversationAndCheckMembershipWithError
+    (errorDescriptionToWai convAccessDenied)
 
 getConversationAndCheckMembershipWithError :: Error -> UserId -> ConvId -> Galley Data.Conversation
 getConversationAndCheckMembershipWithError ex zusr convId = do
-  c <- Data.conversation convId >>= ifNothing convNotFound
+  c <- Data.conversation convId >>= ifNothing (errorDescriptionToWai convNotFound)
   when (DataTypes.isConvDeleted c) $ do
     Data.deleteConversation convId
-    throwM convNotFound
+    throwErrorDescription convNotFound
   unless (zusr `isMember` Data.convLocalMembers c) $
     throwM ex
   return c
@@ -320,14 +323,16 @@ pushConversationEvent conn e users bots = do
 
 verifyReusableCode :: ConversationCode -> Galley DataTypes.Code
 verifyReusableCode convCode = do
-  c <- Data.lookupCode (conversationKey convCode) DataTypes.ReusableCode >>= ifNothing codeNotFound
+  c <-
+    Data.lookupCode (conversationKey convCode) DataTypes.ReusableCode
+      >>= ifNothing (errorDescriptionToWai codeNotFound)
   unless (DataTypes.codeValue c == conversationCode convCode) $
-    throwM codeNotFound
+    throwM (errorDescriptionToWai codeNotFound)
   return c
 
 ensureConversationAccess :: UserId -> ConvId -> Access -> Galley Data.Conversation
 ensureConversationAccess zusr cnv access = do
-  conv <- Data.conversation cnv >>= ifNothing convNotFound
+  conv <- Data.conversation cnv >>= ifNothing (errorDescriptionToWai convNotFound)
   ensureAccess conv access
   zusrMembership <- maybe (pure Nothing) (`Data.teamMember` zusr) (Data.convTeam conv)
   ensureAccessRole (Data.convAccessRole conv) [(zusr, zusrMembership)]
@@ -336,7 +341,7 @@ ensureConversationAccess zusr cnv access = do
 ensureAccess :: Data.Conversation -> Access -> Galley ()
 ensureAccess conv access =
   unless (access `elem` Data.convAccess conv) $
-    throwM convAccessDenied
+    throwErrorDescription convAccessDenied
 
 --------------------------------------------------------------------------------
 -- Federation